requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
boto3==1.35.20
redis==5.1.0
geopy==2.4.1
gpxpy==1.6.2
celery==5.4.0
flower==2.0.1
pillow==10.4.0
python-slugify==8.0.4
anthropic==0.34.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
